+++ farcaster/config.py
@@ -3,7 +3,7 @@ from typing import Any, Dict, Optional
 from farcaster.models import *
 
 FARCASTER_API_BASE_URL = "https://api.warpcast.com/v2/"
-NEYNAR_API_BASE_URL = "https://api.neynar.io/v2/farcaster/"
+NEYNAR_API_BASE_URL = "https://api.neynar.com/v1/farcaster/"
 
 
 class ConfigurationParams(BaseModel):
